+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add: getInitDatetime() event and initDatetime() method. (vistart)
 - Add: isInitDatetime() method for judging whether specified attribute is init date & time. (vistart)
 - Chg: isInitDatetime() considers the null value is init value. (vistart)
+- Add: $initDatetime & $initTimestamp static properties. (vistart);
 
 #### traits/PasswordTrait.php